Charts
Gender_Years: 
Bar chart showing the proportion of male and female athletes participating in the Olympic Games over the years.
This chart illustrates the gender distribution of athletes in the Olympic Games from 1896 to 2016. It provides a visual representation of how the number of male and female athletes has evolved over time. 2024 was the first time the we had the same number of women and men competing in the Olympic games, after 128 years of modern competitions.

woman_sports: 
Heatmap displaying the number of female athletes in each sport over the years.
The heatmap visualizes the participation of female athletes across different sports over time. Each cell represents the number of female athletes in a specific sport for a given year. This heatmap helps identify which sports have seen growth in female participation and how the distribution of female athletes has changed across various disciplines over the years.

heigh_weight:
Scatter plot with a regression curve showing the relationship between the height and weight of athletes.
This scatter plot with a regression curve depicts the correlation between the height and weight of athletes. The regression line helps understand the trend or relationship between these two variables. This visualization can provide insights into body composition patterns among athletes and may reveal any underlying trends or anomalies.